requirements.txt
fastapi==0.100.1
passlib[bcrypt]
pydantic==2.1.1
python-jose[cryptography]
SQLAlchemy==2.0.19
starlette==0.31.0
